+++ src/gui/mod.rs
@@ -1750,7 +1750,17 @@ fn apply_config(ajour: &mut Ajour, config: Config) {
                     })
                     .next()
                 {
-                    a.width = column.width.map_or(Length::Fill, Length::Units);
+                    // Always force "Title" column as Length::Fill
+                    //
+                    // Shouldn't be an issue here, as it was for catalog column fix
+                    // below, but will cover things in case anyone accidently manually
+                    // modifies their config and sets a fixed width on this column.
+                    a.width = if a.key == ColumnKey::Title {
+                        Length::Fill
+                    } else {
+                        column.width.map_or(Length::Fill, Length::Units)
+                    };
+
                     a.hidden = column.hidden;
                     a.order = idx;
                 }
@@ -1807,7 +1817,18 @@ fn apply_config(ajour: &mut Ajour, config: Config) {
                     })
                     .next()
                 {
-                    a.width = column.width.map_or(Length::Fill, Length::Units);
+                    // Always force "Title" column as Length::Fill
+                    //
+                    // An older version of ajour used a different column as the fill
+                    // column and some users have migration issues when updating to
+                    // a newer version, causing NO columns to be set as Fill and
+                    // making resizing columns work incorrectly
+                    a.width = if a.key == CatalogColumnKey::Title {
+                        Length::Fill
+                    } else {
+                        column.width.map_or(Length::Fill, Length::Units)
+                    };
+
                     a.hidden = column.hidden;
                     a.order = idx;
                 }