+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and `Removed`.
 
 - Certain CF addons weren't fingerprinted correctly due to a bug in the fingerprinting
   logic for splitting on lines.
+- Fix a bug where users upgrading from an older version of Ajour might have incorrect
+  behavior when trying to resize a column in the Catalog
 
 ## [0.5.4] - 2020-12-07